mesh size computation

+++ source/blender/blenkernel/BKE_geometry_set.hh
@@ -115,7 +115,7 @@ class GeometryComponent : public ImplicitSharingMixin {
 
   virtual bool is_empty() const;
 
-  virtual int64_t size_in_bytes_approximate() const = 0;
+  virtual size_t size_in_bytes_approximate() const = 0;
 
  private:
   void delete_self() override;
@@ -338,7 +338,7 @@ struct GeometrySet {
    */
   bool is_empty() const;
 
-  int64_t size_in_bytes_approximate() const;
+  size_t size_in_bytes_approximate() const;
 
   /**
    * Returns a read-only mesh or null.
@@ -495,7 +495,7 @@ class MeshComponent : public GeometryComponent {
   Mesh *get_for_write();
 
   bool is_empty() const final;
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
@@ -555,7 +555,7 @@ class PointCloudComponent : public GeometryComponent {
   PointCloud *get_for_write();
 
   bool is_empty() const final;
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
@@ -602,7 +602,7 @@ class CurveComponent : public GeometryComponent {
   Curves *get_for_write();
 
   bool is_empty() const final;
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
@@ -643,7 +643,7 @@ class InstancesComponent : public GeometryComponent {
                GeometryOwnershipType ownership = GeometryOwnershipType::Owned);
 
   bool is_empty() const final;
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
@@ -693,7 +693,7 @@ class VolumeComponent : public GeometryComponent {
    */
   Volume *get_for_write();
 
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
@@ -734,7 +734,7 @@ class GeometryComponentEditData final : public GeometryComponent {
 
   void clear() override;
 
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   /**
    * The first node that does topology changing operations on curves should store the curve point
@@ -779,7 +779,7 @@ class GreasePencilComponent : public GeometryComponent {
   GreasePencil *get_for_write();
 
   bool is_empty() const final;
-  int64_t size_in_bytes_approximate() const final;
+  size_t size_in_bytes_approximate() const final;
 
   bool owns_direct_data() const override;
   void ensure_owns_direct_data() override;
