+++ source/blender/blenkernel/intern/geometry_component_instances.cc
@@ -64,7 +64,7 @@ bool InstancesComponent::is_empty() const
   return true;
 }
 
-int64_t InstancesComponent::size_in_bytes_approximate() const
+size_t InstancesComponent::size_in_bytes_approximate() const
 {
   return 0;
 }
